go to sleep

+++ hooks.c
@@ -6,20 +6,20 @@
 /*   By: tbruha <tbruha@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/10 17:44:40 by tbruha            #+#    #+#             */
-/*   Updated: 2025/02/12 23:35:35 by tbruha           ###   ########.fr       */
+/*   Updated: 2025/02/13 00:40:20 by tbruha           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "include/fractol.h"
 
-void	resize_mgmt(int32_t width, int32_t height, void* param)
+void	resize_mgmt(int32_t width, int32_t height, void *param)
 {
-	t_fractal *fract;
+	t_fractal	*fract;
 
 	fract = param;
 	fract->side_x = width;
 	fract->side_y = height;
-	fract->img = mlx_new_image(fract->mlx_cnct, (uint32_t)width, (uint32_t)height);
+	fract->img = mlx_new_image(fract->mlx_cnct, width, height);
 	rndr_fract(fract);
 }
